Refactor: restore split into QueryAPI and PqlAPI
and reorganize the tests
