+++ pypuppetdb/api/__init__.py
@@ -1,6 +1,7 @@
-from pypuppetdb.api.query import QueryAPI
 from pypuppetdb.api.metrics import MetricsAPI
 from pypuppetdb.api.other import CmdAPI, MetaAPI, ServicesAPI
+from pypuppetdb.api.pql import PqlAPI
+from pypuppetdb.api.query import QueryAPI
 
 
 class API(
@@ -9,5 +10,6 @@ class API(
     CmdAPI,
     MetaAPI,
     ServicesAPI,
+    PqlAPI,
 ):
     pass
